LayoutTests/media/modern-media-controls/media-controls/media-controls-start-button.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>MediaControls</code> showsStartButton property.");

const mediaControls = new MediaControls({ layoutTraits: LayoutTraits.macOS });

debug("Default where showsStartButton = false");
shouldBeFalse("mediaControls.showsStartButton");
shouldBe("mediaControls.children", "[mediaControls.controlsBar]");

debug("");
debug("showsStartButton = true");
mediaControls.showsStartButton = true;
shouldBeTrue("mediaControls.showsStartButton");
shouldBe("mediaControls.children", "[mediaControls.startButton]");

debug("");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
